requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
pdfplumber>=0.10.0
